requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14
drf-spectacular>=0.27
PyYAML>=6.0.1
requests>=2.32.0
